Move the JSON query compiler into beacon-core; delete beacon-query
The JSON query model and compiler (QueryBody/Select/Filter/From/output + the
QueryBody->LogicalPlan builder) move from the beacon-query crate into
beacon-core/src/query. plan_client_query now branches on the inner query: SQL goes
straight to DataFusion's SQL parser (plan_sql_client_query), JSON is compiled by
beacon's own compiler (query::compile_json_query). Both feed execute_statement_plan.

Because the compiler is impls on the query types, the whole model moves with it,
which empties beacon-query -- so the crate is deleted (removed from the workspace
and from beacon-core's deps; cast/geojson/chrono/beacon-arrow-tiff pulled into
beacon-core). beacon-planner's now-dead plan_query/wrap_file_scans/BeaconQueryPlan
(unused since the query path was unified) are removed too, leaving it metrics-only.

explain_client_query now builds its plan through plan_client_query.

Workspace builds; beacon-core (incl. JSON query tests) + beacon-api + beacon-planner
tests pass.
