+++ beacon-core/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod api;
 pub mod parser;
+pub mod query;
 pub mod query_result;
 pub mod runtime;
 mod statement_plan;
