+++ beacon-core/src/query/filter/eq.rs
@@ -6,59 +6,56 @@ use datafusion::{
     prelude::{lit_timestamp_nano, Expr},
 };
 
-use crate::filter::try_coerce_number_to_schema;
+use crate::query::filter::try_coerce_number_to_schema;
 
 use super::{get_column_type, parse_column_name};
 
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize, utoipa::ToSchema)]
 #[serde(untagged)]
-pub enum Lteq {
+pub enum Eq {
     Number {
         #[serde(alias = "for_query_parameter")]
         column: String,
-        #[serde(alias = "max")]
-        lt_eq: f64,
+        eq: f64,
     },
     Timestamp {
         #[serde(alias = "for_query_parameter")]
         column: String,
-        #[serde(alias = "max")]
-        lt_eq: NaiveDateTime,
+        eq: NaiveDateTime,
     },
     String {
         #[serde(alias = "for_query_parameter")]
         column: String,
-        #[serde(alias = "max")]
-        lt_eq: String,
+        eq: String,
     },
 }
 
-impl Lteq {
+impl Eq {
     pub fn parse(
         &self,
         _session_state: &SessionState,
         schema: &Schema,
     ) -> datafusion::error::Result<Expr> {
         match self {
-            Lteq::Number { lt_eq, column } => {
+            Eq::Number { eq, column } => {
                 let column_type = get_column_type(schema, &column);
                 let column = parse_column_name(&column);
 
                 match column_type {
                     Some(dtype) => {
-                        let coerced_lit = try_coerce_number_to_schema(*lt_eq, &dtype);
-                        Ok(column.lt_eq(coerced_lit))
+                        let coerced_lit = try_coerce_number_to_schema(*eq, &dtype);
+                        Ok(column.eq(coerced_lit))
                     }
-                    None => Ok(column.lt_eq(lit(*lt_eq))),
+                    None => Ok(column.eq(lit(*eq))),
                 }
             }
-            Lteq::Timestamp { lt_eq, column } => {
+            Eq::Timestamp { eq, column } => {
                 let column = parse_column_name(&column);
-                Ok(column.lt_eq(lit_timestamp_nano(lt_eq.timestamp_nanos())))
+                Ok(column.eq(lit_timestamp_nano(eq.timestamp_nanos())))
             }
-            Lteq::String { lt_eq, column } => {
+            Eq::String { eq, column } => {
                 let column = parse_column_name(&column);
-                Ok(column.lt_eq(lit(lt_eq)))
+                Ok(column.eq(lit(eq)))
             }
         }
     }
