+++ beacon-core/src/query/filter/geo_json.rs
