+++ beacon-core/src/query/filter/lte.rs
@@ -6,59 +6,59 @@ use datafusion::{
     prelude::{lit_timestamp_nano, Expr},
 };
 
-use crate::filter::try_coerce_number_to_schema;
+use crate::query::filter::try_coerce_number_to_schema;
 
 use super::{get_column_type, parse_column_name};
 
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize, utoipa::ToSchema)]
 #[serde(untagged)]
-pub enum Gteq {
+pub enum Lteq {
     Number {
         #[serde(alias = "for_query_parameter")]
         column: String,
-        #[serde(alias = "min")]
-        gt_eq: f64,
+        #[serde(alias = "max")]
+        lt_eq: f64,
     },
     Timestamp {
         #[serde(alias = "for_query_parameter")]
         column: String,
-        #[serde(alias = "min")]
-        gt_eq: NaiveDateTime,
+        #[serde(alias = "max")]
+        lt_eq: NaiveDateTime,
     },
     String {
         #[serde(alias = "for_query_parameter")]
         column: String,
-        #[serde(alias = "min")]
-        gt_eq: String,
+        #[serde(alias = "max")]
+        lt_eq: String,
     },
 }
 
-impl Gteq {
+impl Lteq {
     pub fn parse(
         &self,
         _session_state: &SessionState,
         schema: &Schema,
     ) -> datafusion::error::Result<Expr> {
         match self {
-            Gteq::Number { gt_eq, column } => {
+            Lteq::Number { lt_eq, column } => {
                 let column_type = get_column_type(schema, &column);
                 let column = parse_column_name(&column);
 
                 match column_type {
                     Some(dtype) => {
-                        let coerced_lit = try_coerce_number_to_schema(*gt_eq, &dtype);
-                        Ok(column.gt_eq(coerced_lit))
+                        let coerced_lit = try_coerce_number_to_schema(*lt_eq, &dtype);
+                        Ok(column.lt_eq(coerced_lit))
                     }
-                    None => Ok(column.gt_eq(lit(*gt_eq))),
+                    None => Ok(column.lt_eq(lit(*lt_eq))),
                 }
             }
-            Gteq::Timestamp { gt_eq, column } => {
+            Lteq::Timestamp { lt_eq, column } => {
                 let column = parse_column_name(&column);
-                Ok(column.gt_eq(lit_timestamp_nano(gt_eq.timestamp_nanos())))
+                Ok(column.lt_eq(lit_timestamp_nano(lt_eq.timestamp_nanos())))
             }
-            Gteq::String { gt_eq, column } => {
+            Lteq::String { lt_eq, column } => {
                 let column = parse_column_name(&column);
-                Ok(column.gt_eq(lit(gt_eq)))
+                Ok(column.lt_eq(lit(lt_eq)))
             }
         }
     }
