+++ beacon-core/src/query/output.rs
