+++ beacon-planner/Cargo.toml
@@ -15,9 +15,3 @@ parking_lot = { workspace = true}
 geojson = "0.24.2"
 uuid = { version = "1.16.0", features = ["serde"] }
 chrono = { workspace = true}
-
-beacon-config = { path = "../beacon-config" }
-beacon-functions = { path = "../beacon-functions" }
-beacon-query = { path = "../beacon-query" }
-beacon-data-lake = { path = "../beacon-data-lake" }
-beacon-formats = { path = "../beacon-formats" }
