+++ beacon-planner/src/lib.rs
@@ -1,7 +1,5 @@
 pub mod metrics;
-pub mod plan;
 
 pub mod prelude {
     pub use crate::metrics::MetricsTracker;
-    pub use crate::plan::*;
 }
