--- beacon-query/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "beacon-query"
-version = "1.7.0"
-edition = "2021"
-
-[dependencies]
-utoipa = { workspace=true}
-serde = { workspace=true}
-tokio = { workspace = true }
-datafusion = { workspace=true}
-tracing = { workspace=true}
-anyhow = { workspace=true}
-tempfile = { workspace=true}
-cast = "0.3.0"
-geojson = "0.24.2"
-chrono = { workspace = true }
-
-beacon-config = { path = "../beacon-config" }
-beacon-functions = { path = "../beacon-functions" }
-beacon-data-lake = { path = "../beacon-data-lake" }
-beacon-formats = { path = "../beacon-formats" }
-beacon-arrow-odv = { path = "../beacon-arrow-odv" }
-beacon-arrow-netcdf = { path = "../beacon-arrow-netcdf" }
-beacon-arrow-tiff = { path = "../beacon-arrow-tiff" }
-beacon-object-storage = { path = "../beacon-object-storage" }
