--- beacon-query/src/parser.rs
@@ -1,144 +1,0 @@
-use beacon_data_lake::{FileManager, TableManager};
-use datafusion::{
-    logical_expr::LogicalPlan,
-    prelude::{SQLOptions, SessionContext},
-};
-
-use crate::{plan::ParsedPlan, InnerQuery, QueryBody};
-
-use super::Query;
-
-pub struct Parser;
-
-impl Parser {
-    pub async fn parse(
-        session: &SessionContext,
-        table_manager: &TableManager,
-        file_manager: &FileManager,
-        query: Query,
-    ) -> anyhow::Result<ParsedPlan> {
-        let datafusion_logical_plan =
-            Self::parse_to_logical_plan(session, table_manager, file_manager, query.inner).await?;
-
-        if let Some(output) = &query.output {
-            let (plan, output_file) = output
-                .parse(session, file_manager, datafusion_logical_plan)
-                .await?;
-
-            Ok(ParsedPlan::new(plan, Some(output_file)))
-        } else {
-            Ok(ParsedPlan::new(datafusion_logical_plan, None))
-        }
-    }
-
-    pub async fn parse_to_logical_plan(
-        session: &SessionContext,
-        table_manager: &TableManager,
-        file_manager: &FileManager,
-        inner_query: InnerQuery,
-    ) -> anyhow::Result<LogicalPlan> {
-        let datafusion_logical_plan = match inner_query {
-            InnerQuery::Sql(sql) => {
-                if beacon_config::CONFIG.sql.enable {
-                    let sql_options = SQLOptions::new()
-                        .with_allow_ddl(false)
-                        .with_allow_dml(false)
-                        .with_allow_statements(false);
-                    let logical_plan = session
-                        .sql_with_options(&sql, sql_options)
-                        .await?
-                        .into_parts()
-                        .1;
-                    logical_plan
-                } else {
-                    // Return an error if SQL queries are not enabled
-                    tracing::warn!("SQL queries are not enabled");
-                    return Err(anyhow::anyhow!("SQL queries are not enabled"));
-                }
-            }
-            InnerQuery::Json(query_body) => {
-                let datafusion_plan =
-                    Self::parse_json_query(query_body, session, table_manager, file_manager).await?;
-                datafusion_plan
-            }
-        };
-
-        Ok(datafusion_logical_plan)
-    }
-
-    pub async fn parse_json_query(
-        query_body: QueryBody,
-        session: &SessionContext,
-        table_manager: &TableManager,
-        file_manager: &FileManager,
-    ) -> anyhow::Result<LogicalPlan> {
-        let mut builder = if beacon_config::CONFIG.sql.enable_pushdown_projection {
-            let mut all_columns = vec![];
-            for select in &query_body.select {
-                let mut select_cols = vec![];
-                select.collect_columns(&mut select_cols);
-                all_columns.extend(select_cols);
-            }
-
-            query_body
-                .from
-                .unwrap_or_default()
-                .init_builder(session, table_manager, file_manager, Some(&all_columns))
-                .await?
-        } else {
-            query_body
-                .from
-                .unwrap_or_default()
-                .init_builder(session, table_manager, file_manager, None)
-                .await?
-        };
-
-        let session_state = session.state();
-
-        builder = builder.project(
-            query_body
-                .select
-                .iter()
-                .map(|s| s.to_expr(&session.state()))
-                .collect::<anyhow::Result<Vec<_>>>()?,
-        )?;
-
-        let df_schema = builder.schema().clone();
-        let schema = df_schema.as_arrow();
-        if let Some(filter) = query_body.filter {
-            builder = builder.filter(filter.parse(&session_state, schema)?)?;
-        }
-
-        if let Some(filters) = query_body.filters {
-            for filter in filters {
-                builder = builder.filter(filter.parse(&session_state, schema)?)?;
-            }
-        }
-
-        if let Some(sort_by) = query_body.sort_by {
-            builder = builder.sort(sort_by.iter().map(|s| s.to_expr()))?;
-        }
-
-        if let Some(distinct) = query_body.distinct {
-            let on_exprs = distinct
-                .on
-                .iter()
-                .map(|s| s.to_expr(&session.state()))
-                .collect::<anyhow::Result<Vec<_>>>()?;
-
-            let select_exprs = distinct
-                .select
-                .iter()
-                .map(|s| s.to_expr(&session.state()))
-                .collect::<anyhow::Result<Vec<_>>>()?;
-
-            builder = builder.distinct_on(on_exprs, select_exprs, None)?;
-        }
-
-        let offset = query_body.offset.unwrap_or(0);
-        builder = builder.limit(offset, query_body.limit)?;
-
-        let plan = builder.build()?;
-        Ok(plan)
-    }
-}
