--- beacon-query/src/util.rs
@@ -1,8 +1,0 @@
-use datafusion::{
-    common::Column,
-    prelude::{col, Expr},
-};
-
-pub(crate) fn parse_column_name(name: &str) -> Expr {
-    col(Column::from_name(name))
-}
